init: Load DLKM modules separately to avoid issues when the user has a custom kernel that breaks the ABI.
Even if the recovery kernel and modules are built together, TWRP will not load the DLKM modules if the boot.img kernel and recovery.img kernel have different ABI versions.

This is very likely to happen when the user has a custom kernel installed, so to avoid issues, we load the DLKM modules separately.

Fixes touchscreen not working in TWRP when a custom kernel is installed.

Signed-off-by: ravindu644 <droidcasts@protonmail.com>